Implement Setting bindings
* move initUserPreferenceSettings to data/stores/settings.ts
* get defaults for userSettings from userPreferenceSettings
* add mergeDefaultStorage to handle when new defaults come in
* add bind:value on settings/+page.svelte to userSettings
* use userSettings for Grid vs. List in BookSelector
* use userSettings for showVerseSelector in BookSelector and
  ChapterSelector
* use userSettings for language
