+++ src/lib/components/ChapterSelector.svelte
@@ -6,12 +6,12 @@ The navbar component.
     import Dropdown from './Dropdown.svelte';
     import SelectGrid from './SelectGrid.svelte';
     import TabsMenu from './TabsMenu.svelte';
-    import { refs, nextRef, s, t, convertStyle } from '$lib/data/stores';
+    import { refs, nextRef, s, t, convertStyle, userSettings } from '$lib/data/stores';
     import { DropdownIcon } from '$lib/icons';
     import { catalog } from '$lib/data/catalog';
     import config from '$lib/data/config';
 
-    const showVerseSelector = config.mainFeatures['show-verse-selector'];
+    const showVerseSelector = $userSettings['verse-selection'];
 
     /**reference to chapter selector so code can use TabsMenu.setActive*/
     let chapterSelector;
