+++ src/lib/data/stores/localization.js
@@ -1,13 +1,11 @@
-import { writable, derived } from 'svelte/store';
-import { setDefaultStorage } from './storage';
+import { derived } from 'svelte/store';
+import { userSettings } from './setting';
 import config from '../config';
 
 /** localization */
 export const languageDefault = config.translationMappings.defaultLang;
-setDefaultStorage('language', languageDefault);
 export const languages = Object.keys(config.interfaceLanguages.writingSystems);
-export const language = writable(localStorage.language);
-language.subscribe(value => localStorage.language = value);
+export const language = derived(userSettings, $userSettings => $userSettings['interface-language']);
 
 export const t = derived(language, $language => {
     return Object.keys(config.translationMappings.mappings).reduce((mappings, key) => {
